working

+++ lab_4/src/pwm.h
@@ -4,7 +4,7 @@
 void initPWMTimer3();
 void initPWMTimer4();
 
-void changeDutyCycle(double dutycycle1,double dutycycle2);
+void changeDutyCycle(float dutycycle1, float dutycycle2);
 
 void analyzeADC();
